third_party/blink/web_tests/http/tests/security/cross-origin-window-open-exception-expected.txt

Cross-origin access to 'window.open' and 'window.opener' should throw a SecurityError.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS frame.contentWindow.open() threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame..
FAIL frame.contentWindow.opener = 1; should throw SecurityError: Failed to set the 'opener' property on 'Window': Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.. Threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame..
PASS successfullyParsed is true

TEST COMPLETE
